+++ .gitignore
@@ -32,3 +32,5 @@ yarn-error.log*
 /**/firebase.service-account.json
 /**/vision.service-account.json
 /**/.runtimeconfig.json
+
+/**/*.jpg
