+++ storage.rules
@@ -7,7 +7,7 @@ service firebase.storage {
 
     match /{userID}/images/{image} {
       allow read: if isAnyAuthenticated(userID);
-      allow write: if isAnyAuthenticated(userID)
+      allow write: if request.auth != null // isAnyAuthenticated(userID)
       && request.resource.size < 5 * 1024 * 1024
       && request.resource.contentType.matches('image/.*');
     }
